Added main.cpp
Create a truth table for all possible 4-bit inputs and their corresponding outputs. 
For each input, the output should be 1 if the input is a prime number and 0 if it is not.
Convert the 4-bit input into a 16-line output. Each output line will correspond to a specific input value.
Connect the output lines of the decoder to the selection inputs of a 16:1 multiplexer. 
The data inputs of the multiplexer should be set to 0, except for the lines corresponding to the prime numbers, which should be set to 1.